pyproject.toml
[project]
name = "TicketCore"
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.0",
    "argon2-cffi>=25.1.0",
    "arq>=0.26.3",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=5.3.1",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[build-system]
requires = ['hatchling']
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = "."
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "polyfactory>=3.2.0",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.2",
    "types-passlib>=1.7.7.20260211",
]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-exclude = [
    "migrations",
]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true

plugins = [
    "pydantic.mypy"
]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = [
    "asyncpg.*",
    "factories.*"
]
ignore_missing_imports = true
